fixing more of the breaks within changes

+++ app/src/main/java/com/seniordesign/wolfpack/quizinator/Database/Rules/Rules.java
@@ -10,7 +10,7 @@ public class Rules {
     private int maxCardCount;
     private long timeLimit;
     private long cardDisplayTime;
-    private int deckId;
+    private long deckId;
     /*
      * Using String because going to use an Array of Strings to limit what choice
      * of Card Type in that class instead of using an enum. We thought this
@@ -19,9 +19,6 @@ public class Rules {
      */
     private String cardTypes;
 
-    /*
-     * @author kuczynskij (10/5/2016)
-     */
     @Override
     public String toString(){
         return "Rules id(" + id + "), maxCardCount(" + maxCardCount +
@@ -30,16 +27,10 @@ public class Rules {
                 deckId + ").";
     }
 
-    /*
-     * @author kuczynskij (10/5/2016)
-     */
     public long getId() {
         return id;
     }
 
-    /*
-     * @author kuczynskij (10/5/2016)
-     */
     public void setId(long id) {
         this.id = id;
     }
@@ -76,7 +67,7 @@ public class Rules {
         this.cardTypes = cardTypes;
     }
 
-    public int getDeckId() { return deckId; }
+    public long getDeckId() { return deckId; }
 
     public void setDeckId(int deckId) { this.deckId = deckId; }
 }
